requirements.txt
# Core FastAPI and Web Server
fastapi==0.116.1
uvicorn==0.35.0
starlette==0.47.1
python-multipart  # For UploadFile support

# PDF Processing
PyPDF2==3.0.1

# Resume Parsing (OCR and Docx support)
pytesseract==0.3.13
pillow==11.3.0
python-docx==1.2.0
lxml==6.0.0

# Google Gemini / Generative AI
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-api-python-client==2.176.0
proto-plus==1.26.1
protobuf==5.29.5
grpcio==1.73.1
grpcio-status==1.71.2
httplib2==0.22.0
googleapis-common-protos==1.70.0

# AWS SDK
boto3==1.39.9
botocore==1.39.9
s3transfer==0.13.1
jmespath==1.0.1

# Utility & Support Libraries
requests==2.32.4
urllib3==2.5.0
idna==3.10
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
click==8.2.1
tqdm==4.67.1
python-dateutil==2.9.0.post0

# Typing and Pydantic
pydantic==2.11.7
pydantic_core==2.33.2
typing_extensions==4.14.1
typing-inspection==0.4.1
annotated-types==0.7.0
anyio==4.9.0
sniffio==1.3.1

# Security / Encoding
pyasn1==0.6.1
pyasn1_modules==0.4.2
rsa==4.9.1

# URL and Template Parsing
uritemplate==4.2.0
pyparsing==3.2.3

# Environment Variable Support
python-dotenv  # <-- Add this for .env file handling (important)

# Async file support (for FastAPI file handling)
aiofiles  # Optional, but recommended for FastAPI when saving files
